AAE-680 Split controllers into users and groups, adding test coverage for converters
