+++ activiti-cloud-services-runtime-bundle/activiti-cloud-services-rest-impl/src/main/java/org/activiti/cloud/services/rest/assemblers/ToCandidateGroupConverter.java
@@ -8,7 +8,7 @@ import org.activiti.cloud.api.process.model.impl.CandidateGroup;;
 public class ToCandidateGroupConverter {
 
     public List<CandidateGroup> from (List<String> users){
-        List<CandidateGroup> list = new ArrayList();
+        List<CandidateGroup> list = new ArrayList<>();
         users.forEach(user -> list.add(new CandidateGroup(user)));
         return list;
     }
