+++ activiti-cloud-services-runtime-bundle/activiti-cloud-services-rest-impl/src/main/java/org/activiti/cloud/services/rest/assemblers/ToCandidateUserConverter.java
@@ -8,7 +8,7 @@ import org.activiti.cloud.api.process.model.impl.CandidateUser;
 public class ToCandidateUserConverter {
 
     public List<CandidateUser> from (List<String> users){
-        List<CandidateUser> list = new ArrayList();
+        List<CandidateUser> list = new ArrayList<>();
         users.forEach(user -> list.add(new CandidateUser(user)));
         return list;
     }
