+++ activiti-cloud-services-runtime-bundle/activiti-cloud-services-rest-impl/src/test/java/org/activiti/cloud/services/rest/assemblers/ToCandidateGroupConverterTest.java
@@ -1,0 +1,34 @@
+package org.activiti.cloud.services.rest.assemblers;
+
+import java.util.ArrayList;
+import java.util.List;
+
+import org.activiti.cloud.api.process.model.impl.CandidateGroup;
+import org.junit.Before;
+import org.junit.Test;
+
+import static org.assertj.core.api.Assertions.*;
+
+public class ToCandidateGroupConverterTest {
+
+    private ToCandidateGroupConverter toCandidateGroupConverter;
+
+    @Before
+    public void setUp(){
+        toCandidateGroupConverter = new ToCandidateGroupConverter();
+    }
+
+    @Test
+    public void shouldConvertStringGroupsToCanidateGroups(){
+        //given
+        String group = "user1";
+        List<String> userList = new ArrayList<>();
+        userList.add(group);
+        //when
+        List<CandidateGroup> convertedGroupList = toCandidateGroupConverter.from(userList);
+        //then
+        assertThat(convertedGroupList.get(0)).isInstanceOf(CandidateGroup.class);
+        assertThat(convertedGroupList.get(0).getGroup()).isEqualTo(group);
+    }
+
+}
