+++ activiti-cloud-services-runtime-bundle/activiti-cloud-services-rest-impl/src/test/java/org/activiti/cloud/services/rest/controllers/CandidateGroupAdminControllerImplIT.java
@@ -3,13 +3,9 @@ package org.activiti.cloud.services.rest.controllers;
 import java.util.Arrays;
 import java.util.List;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
 import org.activiti.api.runtime.conf.impl.CommonModelAutoConfiguration;
-import org.activiti.api.runtime.shared.query.Page;
-import org.activiti.api.runtime.shared.security.SecurityManager;
 import org.activiti.api.task.conf.impl.TaskModelAutoConfiguration;
-import org.activiti.api.task.model.Task;
-import org.activiti.api.task.runtime.TaskRuntime;
+import org.activiti.api.task.runtime.TaskAdminRuntime;
 import org.activiti.cloud.alfresco.config.AlfrescoWebAutoConfiguration;
 import org.activiti.cloud.services.core.pageable.SpringPageConverter;
 import org.activiti.cloud.services.events.ProcessEngineChannels;
@@ -23,7 +19,6 @@ import org.activiti.spring.process.conf.ProcessExtensionsAutoConfiguration;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.mockito.Mock;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.restdocs.AutoConfigureRestDocs;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
@@ -41,15 +36,14 @@ import org.springframework.test.web.servlet.MvcResult;
 import static net.javacrumbs.jsonunit.fluent.JsonFluentAssert.assertThatJson;
 import static org.activiti.alfresco.rest.docs.AlfrescoDocumentation.resourcesResponseFields;
 import static org.activiti.alfresco.rest.docs.HALDocumentation.unpagedCandidateGroups;
-import static org.activiti.alfresco.rest.docs.HALDocumentation.unpagedCandidateUsers;
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.mockito.Mockito.when;
+import static org.assertj.core.api.Assertions.*;
+import static org.mockito.Mockito.*;
 import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.document;
 import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 @RunWith(SpringRunner.class)
-@WebMvcTest(controllers = CandidateControllerImpl.class, secure = true)
+@WebMvcTest(controllers = CandidateGroupAdminControllerImpl.class, secure = true)
 @EnableSpringDataWebSupport
 @AutoConfigureMockMvc(secure = false)
 @AutoConfigureRestDocs(outputDir = "target/snippets")
@@ -61,100 +55,48 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
         ProcessExtensionsAutoConfiguration.class,
         ServicesRestWebMvcAutoConfiguration.class,
         AlfrescoWebAutoConfiguration.class})
-public class CandidateControllerImplIT {
+public class CandidateGroupAdminControllerImplIT {
 
-    private static final String DOCUMENTATION_IDENTIFIER = "task";
+    private static final String DOCUMENTATION_IDENTIFIER = "candidate-group-admin";
+
+    private static final String DOCUMENTATION_IDENTIFIER_ALFRESCO = "candidate-group-admin-alfresco";
 
     @Autowired
     private MockMvc mockMvc;
 
-    @Autowired
-    private ObjectMapper mapper;
+    @MockBean
+    private TaskAdminRuntime taskAdminRuntime;
 
     @MockBean
     private RepositoryService repositoryService;
 
-    @MockBean
-    private SecurityManager securityManager;
-
-    @MockBean
-    private TaskRuntime taskRuntime;
-
     @SpyBean
-    private SpringPageConverter springPageConverter;
+    private SpringPageConverter pageConverter;
 
     @MockBean
     private ProcessEngineChannels processEngineChannels;
-
-    @Mock
-    private Page<Task> taskPage;
 
     @MockBean
     private CloudProcessDeployedProducer processDeployedProducer;
 
     @Before
     public void setUp() {
-        assertThat(springPageConverter).isNotNull();
+        assertThat(pageConverter).isNotNull();
         assertThat(processEngineChannels).isNotNull();
         assertThat(processDeployedProducer).isNotNull();
     }
-
-    @Test
-    public void getUserCandidatesShouldUseAlfrescoGuidelineWhenMediaTypeIsApplicationJson() throws Exception {
-
-        List<String> stringList = Arrays.asList("hruser",
-                                                "testuser");
-        when(taskRuntime.userCandidates("1")).thenReturn(stringList);
-
-        MvcResult result = this.mockMvc.perform(get("/v1/tasks/{taskId}/candidate-users",
-                                                    1).accept(MediaType.APPLICATION_JSON))
-                .andExpect(status().isOk())
-                .andDo(document(DOCUMENTATION_IDENTIFIER + "/list",
-                                resourcesResponseFields()))
-                .andReturn();
-
-        assertThatJson(result.getResponse().getContentAsString())
-                .node("list.entries[0].entry.user")
-                .isEqualTo("hruser");
-        assertThatJson(result.getResponse().getContentAsString())
-                .node("list.entries[1].entry.user")
-                .isEqualTo("testuser");
-    }
-
-    @Test
-    public void getUserCandidatesShouldHaveProperHALFormat() throws Exception {
-
-        List<String> stringList = Arrays.asList("hruser",
-                                                "testuser");
-        when(taskRuntime.userCandidates("1")).thenReturn(stringList);
-
-        MvcResult result = this.mockMvc.perform(get("/v1/tasks/{taskId}/candidate-users",
-                                                    1).accept(MediaTypes.HAL_JSON_VALUE))
-                .andExpect(status().isOk())
-                .andDo(document(DOCUMENTATION_IDENTIFIER + "/list",
-                                unpagedCandidateUsers()))
-                .andReturn();
-
-        assertThatJson(result.getResponse().getContentAsString())
-                .node("_embedded.candidateUsers[0].user")
-                .isEqualTo("hruser");
-        assertThatJson(result.getResponse().getContentAsString())
-                .node("_embedded.candidateUsers[1].user")
-                .isEqualTo("testuser");
-    }
-
 
     @Test
     public void getGroupCandidatesShouldUseAlfrescoGuidelineWhenMediaTypeIsApplicationJson() throws Exception {
 
         List<String> stringList = Arrays.asList("hrgroup",
                                                 "testgroup");
-        when(taskRuntime.groupCandidates("1")).thenReturn(stringList);
+        when(taskAdminRuntime.groupCandidates("1")).thenReturn(stringList);
 
-        MvcResult result = this.mockMvc.perform(get("/v1/tasks/{taskId}/candidate-groups",
+        MvcResult result = this.mockMvc.perform(get("/admin/v1/tasks/{taskId}/candidate-groups",
                                                     1).accept(MediaType.APPLICATION_JSON))
                 .andExpect(status().isOk())
-                .andDo(document(DOCUMENTATION_IDENTIFIER + "/list",
+                .andDo(document(DOCUMENTATION_IDENTIFIER_ALFRESCO + "/list",
                                 resourcesResponseFields()))
                 .andReturn();
 
@@ -171,9 +113,9 @@ public class CandidateControllerImplIT {
 
         List<String> stringList = Arrays.asList("hrgroup",
                                                 "testgroup");
-        when(taskRuntime.groupCandidates("1")).thenReturn(stringList);
+        when(taskAdminRuntime.groupCandidates("1")).thenReturn(stringList);
 
-        MvcResult result = this.mockMvc.perform(get("/v1/tasks/{taskId}/candidate-groups",
+        MvcResult result = this.mockMvc.perform(get("/admin/v1/tasks/{taskId}/candidate-groups",
                                                     1).accept(MediaTypes.HAL_JSON_VALUE))
                 .andExpect(status().isOk())
                 .andDo(document(DOCUMENTATION_IDENTIFIER + "/list",
@@ -187,7 +129,4 @@ public class CandidateControllerImplIT {
                 .node("_embedded.candidateGroups[1].group")
                 .isEqualTo("testgroup");
     }
-
-
-
 }
