+++ activiti-cloud-services-runtime-bundle/activiti-cloud-services-rest-impl/src/test/java/org/activiti/cloud/services/rest/controllers/CandidateGroupControllerImplIT.java
@@ -1,0 +1,147 @@
+package org.activiti.cloud.services.rest.controllers;
+
+import java.util.Arrays;
+import java.util.List;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import org.activiti.api.runtime.conf.impl.CommonModelAutoConfiguration;
+import org.activiti.api.runtime.shared.query.Page;
+import org.activiti.api.runtime.shared.security.SecurityManager;
+import org.activiti.api.task.conf.impl.TaskModelAutoConfiguration;
+import org.activiti.api.task.model.Task;
+import org.activiti.api.task.runtime.TaskRuntime;
+import org.activiti.cloud.alfresco.config.AlfrescoWebAutoConfiguration;
+import org.activiti.cloud.services.core.pageable.SpringPageConverter;
+import org.activiti.cloud.services.events.ProcessEngineChannels;
+import org.activiti.cloud.services.events.configuration.CloudEventsAutoConfiguration;
+import org.activiti.cloud.services.events.configuration.RuntimeBundleProperties;
+import org.activiti.cloud.services.events.listeners.CloudProcessDeployedProducer;
+import org.activiti.cloud.services.rest.conf.ServicesRestWebMvcAutoConfiguration;
+import org.activiti.common.util.conf.ActivitiCoreCommonUtilAutoConfiguration;
+import org.activiti.engine.RepositoryService;
+import org.activiti.spring.process.conf.ProcessExtensionsAutoConfiguration;
+import org.junit.Before;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.mockito.Mock;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.restdocs.AutoConfigureRestDocs;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
+import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.boot.test.mock.mockito.SpyBean;
+import org.springframework.context.annotation.Import;
+import org.springframework.data.web.config.EnableSpringDataWebSupport;
+import org.springframework.hateoas.MediaTypes;
+import org.springframework.http.MediaType;
+import org.springframework.test.context.junit4.SpringRunner;
+import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.MvcResult;
+
+import static net.javacrumbs.jsonunit.fluent.JsonFluentAssert.assertThatJson;
+import static org.activiti.alfresco.rest.docs.AlfrescoDocumentation.resourcesResponseFields;
+import static org.activiti.alfresco.rest.docs.HALDocumentation.unpagedCandidateGroups;
+import static org.assertj.core.api.Assertions.*;
+import static org.mockito.Mockito.*;
+import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.document;
+import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.get;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+
+@RunWith(SpringRunner.class)
+@WebMvcTest(controllers = CandidateGroupControllerImpl.class, secure = true)
+@EnableSpringDataWebSupport
+@AutoConfigureMockMvc(secure = false)
+@AutoConfigureRestDocs(outputDir = "target/snippets")
+@Import({CommonModelAutoConfiguration.class,
+        TaskModelAutoConfiguration.class,
+        RuntimeBundleProperties.class,
+        CloudEventsAutoConfiguration.class,
+        ActivitiCoreCommonUtilAutoConfiguration.class,
+        ProcessExtensionsAutoConfiguration.class,
+        ServicesRestWebMvcAutoConfiguration.class,
+        AlfrescoWebAutoConfiguration.class})
+public class CandidateGroupControllerImplIT {
+
+    private static final String DOCUMENTATION_IDENTIFIER = "candidate-group";
+
+    private static final String DOCUMENTATION_IDENTIFIER_ALFRESCO = "candidate-group-alfresco";
+
+    @Autowired
+    private MockMvc mockMvc;
+
+    @Autowired
+    private ObjectMapper mapper;
+
+    @MockBean
+    private RepositoryService repositoryService;
+
+    @MockBean
+    private SecurityManager securityManager;
+
+    @MockBean
+    private TaskRuntime taskRuntime;
+
+    @SpyBean
+    private SpringPageConverter springPageConverter;
+
+    @MockBean
+    private ProcessEngineChannels processEngineChannels;
+
+    @Mock
+    private Page<Task> taskPage;
+
+    @MockBean
+    private CloudProcessDeployedProducer processDeployedProducer;
+
+    @Before
+    public void setUp() {
+        assertThat(springPageConverter).isNotNull();
+        assertThat(processEngineChannels).isNotNull();
+        assertThat(processDeployedProducer).isNotNull();
+    }
+
+    @Test
+    public void getGroupCandidatesShouldUseAlfrescoGuidelineWhenMediaTypeIsApplicationJson() throws Exception {
+
+        List<String> stringList = Arrays.asList("hrgroup",
+                                                "testgroup");
+        when(taskRuntime.groupCandidates("1")).thenReturn(stringList);
+
+        MvcResult result = this.mockMvc.perform(get("/v1/tasks/{taskId}/candidate-groups",
+                                                    1).accept(MediaType.APPLICATION_JSON))
+                .andExpect(status().isOk())
+                .andDo(document(DOCUMENTATION_IDENTIFIER_ALFRESCO + "/list",
+                                resourcesResponseFields()))
+                .andReturn();
+
+        assertThatJson(result.getResponse().getContentAsString())
+                .node("list.entries[0].entry.group")
+                .isEqualTo("hrgroup");
+        assertThatJson(result.getResponse().getContentAsString())
+                .node("list.entries[1].entry.group")
+                .isEqualTo("testgroup");
+    }
+
+    @Test
+    public void getGroupCandidatesShouldHaveProperHALFormat() throws Exception {
+
+        List<String> stringList = Arrays.asList("hrgroup",
+                                                "testgroup");
+        when(taskRuntime.groupCandidates("1")).thenReturn(stringList);
+
+        MvcResult result = this.mockMvc.perform(get("/v1/tasks/{taskId}/candidate-groups",
+                                                    1).accept(MediaTypes.HAL_JSON_VALUE))
+                .andExpect(status().isOk())
+                .andDo(document(DOCUMENTATION_IDENTIFIER + "/list",
+                                unpagedCandidateGroups()))
+                .andReturn();
+
+        assertThatJson(result.getResponse().getContentAsString())
+                .node("_embedded.candidateGroups[0].group")
+                .isEqualTo("hrgroup");
+        assertThatJson(result.getResponse().getContentAsString())
+                .node("_embedded.candidateGroups[1].group")
+                .isEqualTo("testgroup");
+    }
+
+}
